# src/main/resources/影院订票系统.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "影院订票系统",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/movies/search": {
      "get": {
        "summary": "搜索电影",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "搜索关键词(默认为空)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "页码(默认1)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量(默认10) ",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "string"
                    },
                    "movies": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "poster_url": {
                            "type": "string"
                          },
                          "rating": {
                            "type": "number"
                          },
                          "status": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "status"
                        ]
                      }
                    }
                  },
                  "required": [
                    "total",
                    "movies"
                  ]
                },
                "example": {
                  "total": "Integer, 总数",
                  "movies": [
                    {
                      "id": "Long, 电影ID",
                      "title": "String, 电影标题",
                      "poster_url": "String, 海报URL",
                      "rating": "Double, 评分",
                      "status": "UPCOMING"
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/movies/{movieId}": {
      "get": {
        "summary": "获取电影详情",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "description": "电影ID  ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number"
                    },
                    "title": {
                      "type": "string"
                    },
                    "director": {
                      "type": "string"
                    },
                    "actors": {
                      "type": "string"
                    },
                    "duration": {
                      "type": "integer"
                    },
                    "release_date": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "poster_url": {
                      "type": "string"
                    },
                    "trailer_url": {
                      "type": "string"
                    },
                    "rating": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "director",
                    "actors",
                    "duration",
                    "release_date",
                    "description",
                    "poster_url",
                    "trailer_url",
                    "rating",
                    "status"
                  ]
                },
                "example": {
                  "id": "Long",
                  "title": "String",
                  "director": "String",
                  "actors": "String",
                  "duration": "Integer",
                  "release_date": "Date",
                  "description": "String",
                  "poster_url": "String",
                  "trailer_url": "String",
                  "rating": "Double",
                  "status": "String"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/movies/{movieId}": {
      "put": {
        "summary": "更新电影信息",
        "description": "修改电影基本信息或状态（后台管理）",
        "tags": [
          "电影",
          "后台"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Movie"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        }
      },
      "delete": {
        "summary": "删除电影",
        "description": "永久删除电影（后台管理）",
        "tags": [
          "电影",
          "后台"
        ],
        "parameters": [
          {
            "name": "movieId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        }
      }
    },
...
    "/admin/movies/today": {
      "post": {
        "summary": "设置今日上映电影",
        "description": "管理员设置今日上映电影，body为电影ID数组。",
        "tags": [
          "电影",
          "后台"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer"
                }
              },
              "example": [
                1,
                5,
                8,
                12
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "设置成功"
          }
        }
      },
      "get": {
        "summary": "获取今日上映电影ID列表",
        "description": "管理员获取当前今日上映电影ID数组。",
        "tags": [
          "电影",
          "后台"
        ],
        "responses": {
          "200": {
            "description": "返回今日上映电影ID数组",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                "example": [
                  1,
                  5,
                  8,
                  12
                ]
              }
            }
          }
        }
      }
    },
    "/api/movies/today": {
      "get": {
        "summary": "获取今日上映电影详情",
        "description": "分页获取管理员设置的今日上映电影详情列表。",
        "tags": [
          "电影"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码(默认1)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页数量(默认10)",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "返回今日上映电影分页数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MovieSearchResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "电影ID"
          },
          "title": {
            "type": "string",
            "description": "电影标题"
          },
          "director": {
            "type": "string",
            "description": "导演"
          },
          "actors": {
            "type": "string",
            "description": "演员列表"
          },
          "duration": {
            "type": "integer",
            "description": "时长(分钟)"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "上映日期"
          },
          "description": {
            "type": "string",
            "description": "剧情简介"
          },
          "poster_url": {
            "type": "string",
            "description": "海报URL"
          },
          "trailer_url": {
            "type": "string",
            "description": "预告片URL"
          },
          "rating": {
            "type": "number",
            "format": "float",
            "description": "评分"
          },
          "status": {
            "type": "string",
            "enum": [
              "UPCOMING",
              "SHOWING",
              "OFF"
            ],
            "description": "上映状态"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}